feature: add from operator, split and fix tests
