+++ src/operators/index.ts
@@ -1,2 +1,4 @@
 export * from './predicates';
 export * from './transform';
+export * from './side-effects';
+export * from './static';
