+++ src/operators/static/creation/from.ts
@@ -1,7 +1,5 @@
 import { Observable } from "../../../observable/observable";
 
 export const from = function<T = never>(values: T[]): Observable<T> {
-    const obs$ = new Observable(...values);
-    obs$.close();
-    return obs$;
+    return new Observable(...values);
 }
