+++ src/operators/static/creation/index.ts
@@ -1,0 +1,5 @@
+export * from './from';
+export * from './mutable';
+export * from './of';
+export * from './on-event';
+export * from './pre-process';
