+++ src/operators/static/creation/mutable.ts
@@ -1,0 +1,5 @@
+import { MutableObservable } from "../../../observable/mutable-observable";
+
+export const mutable = function<T = never>(...values: T[]): MutableObservable<T> {
+    return new MutableObservable(...values);
+}
