+++ src/operators/static/creation/of.ts
@@ -1,7 +1,5 @@
 import { Observable } from "../../../observable/observable";
 
-export const of = function<T = never>(value: T): Observable<T> {
-    const obs$ = new Observable(value);
-    obs$.close();
-    return obs$;
+export const of = function<T = never>(...values: T[]): Observable<T> {
+    return new Observable(...values);
 }
