+++ src/operators/static/creation/pre-process.ts
@@ -1,8 +1,8 @@
-import { Observable } from "../../../observable/observable";
 import { OperationResult, OperatorFunction } from "../../../models/operator";
+import { MutableObservable } from "../../../observable/mutable-observable";
 
-export function preProcess<T = never>(...operations: OperatorFunction<any, OperationResult<T>>[]): Observable<T> {
-    const obs$ = new Observable<T>();
+export function preProcess<T = never>(...operations: OperatorFunction<any, OperationResult<T>>[]): MutableObservable<T> {
+    const obs$ = new MutableObservable<T>();
     (obs$ as any)._preProcessOperations.push(...operations);
     return obs$;
 }
